Password sign in refactoring
