+++ app/views/layouts/application.html.haml
@@ -19,7 +19,7 @@
             = nav_link t('landing.navigation.home'), root_landing_url, subdomain: 'www', id: 'home'
             = nav_link t('landing.navigation.bootcamp'), bootcamp_root_url, subdomain: 'bootcamp', id: 'bootcamp'
             = link_to t('landing.navigation.dashboard'), dashboard_root_url, class: 'nav-link' if policy(:dashboard).allowed?
-            = link_to t('landing.navigation.logout'), bootcamp_logout_url, method: :post, class: 'nav-link' if current_user
+            = link_to t('landing.navigation.logout'), logout_url, method: :delete, class: 'nav-link' if current_user
 
       %main.inner.cover{:role => "main"}
         - if flash.alert
