--- app/views/web/bootcamp/user_sessions/new.html.haml
@@ -1,4 +1,0 @@
-= simple_form_for :bootcamp_login, url: bootcamp_user_sessions_url do |f|
-  = f.input :email
-  = f.input :password
-  = f.submit
