+++ app/views/web/idea/sessions/new.html.haml
@@ -1,0 +1,4 @@
+= simple_form_for :idea_login, url: idea_sessions_url do |f|
+  = f.input :email
+  = f.input :password
+  = f.button :submit, t('idea.sessions.submit'), type: :submit, class: 'btn btn-lg btn-secondary'
