+++ config/locales/idea/en.yml
@@ -1,0 +1,4 @@
+en:
+  idea:
+    sessions:
+      submit: Submit
